Added bracet in fun call args on line 64

+++ src/OnUserUpdate.cpp
@@ -64,7 +64,7 @@ bool GameJam::OnUserUpdate(float fElapsedTime)
         if (UIManager.GetPRogressBar() > 0.0f)
         {
             
-            FillRect({GetMouseX() - (int)UIManager.GetMouseText().size() * CHAR_SIZE_ON_SCREEN / 2, GetMouseY() + PROGRESS_BARR_OFFSET},{ static_cast<int>(UIManager.GetPRogressBar() * (float)UIManager.GetMouseText().size() * (float)CHAR_SIZE_ON_SCREEN), PROGRESS_BARR_W}, olc::RED);
+            FillRect({GetMouseX() - (int)UIManager.GetMouseText().size() * CHAR_SIZE_ON_SCREEN / 2, GetMouseY() + PROGRESS_BARR_OFFSET},{ static_cast<int>(UIManager.GetPRogressBar() * UIManager.GetMouseText().size() * CHAR_SIZE_ON_SCREEN), PROGRESS_BARR_W}, olc::RED);
             DrawRect({GetMouseX() - (int)UIManager.GetMouseText().size() * CHAR_SIZE_ON_SCREEN / 2, GetMouseY() + PROGRESS_BARR_OFFSET}, {static_cast<int>( UIManager.GetMouseText().size() * CHAR_SIZE_ON_SCREEN), PROGRESS_BARR_W}, olc::BLACK);
         }
         float ClockScale = 4.0f;
